move repo and ssh key configuration out of the router into it's own file
