+++ packages/@tinacms/api-git/src/configure.test.ts
@@ -1,0 +1,123 @@
+/**
+
+Copyright 2019 Forestry.io Inc
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+
+import { updateRemoteToSSH } from './configure'
+jest.mock('./open-repo')
+
+describe('updateRemoteToSSH', () => {
+  let mock
+
+  let repo: any
+
+  describe('without a remote', () => {
+    beforeEach(() => {
+      repo = {
+        getRemotes: jest.fn().mockImplementation(() => {
+          return Promise.resolve([])
+        }),
+        removeRemote: jest.fn(),
+        addRemote: jest.fn(),
+      }
+      mock = require('./open-repo').openRepo
+      mock.mockImplementation(() => {
+        return repo
+      })
+    })
+
+    it('does not throw an error', async () => {
+      expect(async () => {
+        await updateRemoteToSSH('./')
+      }).not.toThrowError()
+    })
+
+    it('does not tryto remove a remote', async () => {
+      await updateRemoteToSSH('./')
+      expect(repo.removeRemote).not.toHaveBeenCalledWith(
+        'origin',
+        'git@github.com:tinacms/tunacms.git'
+      )
+    })
+
+    it('does not try to add a remote', async () => {
+      await updateRemoteToSSH('./')
+      expect(repo.addRemote).not.toHaveBeenCalledWith(
+        'origin',
+        'git@github.com:tinacms/tunacms.git'
+      )
+    })
+  })
+
+  describe('with http remote', () => {
+    beforeEach(() => {
+      repo = {
+        getRemotes: jest.fn().mockImplementation(() => {
+          return Promise.resolve([
+            {
+              name: 'origin',
+              refs: {
+                push: 'https://github.com/tinacms/tunacms.git',
+              },
+            },
+          ])
+        }),
+        removeRemote: jest.fn(),
+        addRemote: jest.fn(),
+      }
+      mock = require('./open-repo').openRepo
+      mock.mockImplementation(() => {
+        return repo
+      })
+    })
+
+    test('should replace remote', async () => {
+      await updateRemoteToSSH('./')
+      expect(repo.addRemote).toHaveBeenCalledWith(
+        'origin',
+        'git@github.com:tinacms/tunacms.git'
+      )
+    })
+  })
+
+  describe('with ssh remote', () => {
+    beforeEach(() => {
+      repo = {
+        getRemotes: jest.fn().mockImplementation(() => {
+          return Promise.resolve([
+            {
+              name: 'origin',
+              refs: {
+                push: 'git@github.com:tinacms/tunacms.git',
+              },
+            },
+          ])
+        }),
+        removeRemote: jest.fn(),
+        addRemote: jest.fn(),
+      }
+      mock = require('./open-repo').openRepo
+      mock.mockImplementation(() => {
+        return repo
+      })
+    })
+
+    test('should replace remote', async () => {
+      await updateRemoteToSSH('./')
+      expect(repo.addRemote).not.toHaveBeenCalled()
+    })
+  })
+})
