+++ packages/@tinacms/api-git/src/index.ts
@@ -18,3 +18,4 @@ limitations under the License.
 
 export * from './router'
 export * from './server'
+export * from './configure'
